thay doi

+++ src/main/java/trasua/Entity/giohang.java
@@ -1,8 +1,8 @@
 package trasua.Entity;
 
 public class giohang {
-	private int quanty;
-	private float total;
+	private long quanty;
+	private double total;
 	private sanpham product;
 	
 	
@@ -10,23 +10,23 @@ public class giohang {
 	public giohang() {
 	}
 	
-	public giohang(int quanty, float total, sanpham product) {
+	public giohang(long quanty, double total, sanpham product) {
 		super();
 		this.quanty = quanty;
 		this.total = total;
 		this.product = product;
 	}
 	
-	public int getQuanty() {
+	public long getQuanty() {
 		return quanty;
 	}
-	public void setQuanty(int quanty) {
+	public void setQuanty(long quanty) {
 		this.quanty = quanty;
 	}
-	public float getTotal() {
+	public double getTotal() {
 		return total;
 	}
-	public void setTotal(float total) {
+	public void setTotal(double total) {
 		this.total = total;
 	}
 	public sanpham getProduct() {
